# python-flask/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Ytel API version 3"
  version: "3.12"
  title: "Ytel API V3"
host: "api.ytel.com"
basePath: "/api/v3"
schemes:
- "https"
consumes:
- "application/x-www-form-urlencoded"
produces:
- "application/json"
paths:
...
  /incomingphone/availablenumber.json:
    post:
      tags:
      - "Phone Number"
      summary: "Available Numbers"
      description: "Retrieve a list of available phone numbers that can be purchased\
        \ and used for your Ytel account."
      operationId: "incomingphone_availablenumber"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "numbertype"
        in: "formData"
        description: "Number type either SMS,Voice or all"
        required: true
        type: "string"
        enum:
        - "all"
        - "voice"
        - "sms"
        x-enum-elements:
        - name: "ALL"
          description: "Get all phone Number type"
        - name: "Voice"
          description: "Get only voice number type"
        - name: "SMS"
          description: "Get Only SMs number type"
      - name: "areacode"
        in: "formData"
        description: "Specifies the area code for the returned list of available numbers.\
          \ Only available for North American numbers."
        required: true
        type: "string"
      - name: "pagesize"
        in: "formData"
        description: "The count of objects to return."
        required: false
        type: "integer"
        default: 10
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.phone_number_controller"
  /incomingphone/massreleasenumber.json:
    post:
      tags:
      - "Phone Number"
      summary: "Bulk Release"
      description: "Remove a purchased Ytel number from your account."
      operationId: "incomingphone_massreleasenumber"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "PhoneNumber"
        in: "formData"
        description: "A valid Ytel comma separated numbers (E.164 format)."
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.phone_number_controller"
...
  /incomingphone/viewnumber.json:
    post:
      tags:
      - "Phone Number"
      summary: "View Details"
      description: "Retrieve the details for a phone number by its number."
      operationId: "incomingphone_viewnumber"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "PhoneNumber"
        in: "formData"
        description: "A valid Ytel 10-digit phone number (E.164 format)."
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.phone_number_controller"
  /incomingphone/releasenumber.{ResponseType}:
    post:
      tags:
      - "Phone Number"
      summary: "Release Number"
      description: "Remove a purchased Ytel number from your account."
      operationId: "incomingphone_releasenumber_by_response_type_post"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "PhoneNumber"
        in: "formData"
        description: "A valid 10-digit Ytel number (E.164 format)."
        required: true
        type: "string"
      - name: "ResponseType"
        in: "path"
        description: "Response type format xml or json"
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.phone_number_controller"
  /incomingphone/buynumber.json:
    post:
      tags:
      - "Phone Number"
      summary: "Purchase Number"
      description: "Purchase a phone number to be used with your Ytel account"
      operationId: "incomingphone_buynumber"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "PhoneNumber"
        in: "formData"
        description: "A valid 10-digit Ytel number (E.164 format)."
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.phone_number_controller"
...
  /incomingphone/updatenumber.json:
    post:
      tags:
      - "Phone Number"
      summary: "Update Number"
      description: "Update properties for a Ytel number that has been purchased for\
        \ your account. Refer to the parameters list for the list of properties that\
        \ can be updated."
      operationId: "incomingphone_updatenumber"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "PhoneNumber"
        in: "formData"
        description: "A valid Ytel number (E.164 format)."
        required: true
        type: "string"
      - name: "VoiceUrl"
        in: "formData"
        description: "URL requested once the call connects"
        required: true
        type: "string"
      - name: "FriendlyName"
        in: "formData"
        description: "Phone number friendly name description"
        required: false
        type: "string"
      - name: "VoiceMethod"
        in: "formData"
        description: "Post or Get"
        required: false
        type: "string"
      - name: "VoiceFallbackUrl"
        in: "formData"
        description: "URL requested if the voice URL is not available"
        required: false
        type: "string"
      - name: "VoiceFallbackMethod"
        in: "formData"
        description: "Post or Get"
        required: false
        type: "string"
      - name: "HangupCallback"
        in: "formData"
        description: "callback url after a hangup occurs"
        required: false
        type: "string"
      - name: "HangupCallbackMethod"
        in: "formData"
        description: "Post or Get"
        required: false
        type: "string"
      - name: "HeartbeatUrl"
        in: "formData"
        description: "URL requested once the call connects"
        required: false
        type: "string"
      - name: "HeartbeatMethod"
        in: "formData"
        description: "URL that can be requested every 60 seconds during the call to\
          \ notify of elapsed time"
        required: false
        type: "string"
      - name: "SmsUrl"
        in: "formData"
        description: "URL requested when an SMS is received"
        required: false
        type: "string"
      - name: "SmsMethod"
        in: "formData"
        description: "Post or Get"
        required: false
        type: "string"
      - name: "SmsFallbackUrl"
        in: "formData"
        description: "URL used if any errors occur during execution of InboundXML\
          \ from an SMS or at initial request of the SmsUrl."
        required: false
        type: "string"
      - name: "SmsFallbackMethod"
        in: "formData"
        description: "The HTTP method Ytel will use when URL requested if the SmsUrl\
          \ is not available."
        required: false
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.phone_number_controller"
...
  /incomingphone/transferphonenumbers.json:
    post:
      tags:
      - "Phone Number"
      summary: "Move Number"
      description: "Transfer phone number that has been purchased for from one account\
        \ to another account."
      operationId: "incomingphone_transferphonenumbers"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "phonenumber"
        in: "formData"
        description: "A valid 10-digit Ytel number (E.164 format)."
        required: true
        type: "string"
      - name: "fromaccountsid"
        in: "formData"
        description: "A specific Accountsid from where Number is getting transfer."
        required: true
        type: "string"
      - name: "toaccountsid"
        in: "formData"
        description: "A specific Accountsid to which Number is getting transfer."
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.phone_number_controller"
...
  /incomingphone/listnumber.json:
    post:
      tags:
      - "Phone Number"
      summary: "List Numbers"
      description: "Retrieve a list of purchased phones numbers associated with your\
        \ Ytel account."
      operationId: "incomingphone_listnumber"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "Page"
        in: "formData"
        description: "Which page of the overall response will be returned. Page indexing\
          \ starts at 1."
        required: false
        type: "integer"
        default: 1
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "PageSize"
        in: "formData"
        description: "The page count to retrieve from the total results in the collection.\
          \ Page indexing starts at 1."
        required: false
        type: "integer"
        default: 10
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "NumberType"
        in: "formData"
        description: "The capability supported by the number.Number type either SMS,Voice\
          \ or all"
        required: false
        type: "string"
        enum:
        - "all"
        - "voice"
        - "sms"
        x-enum-elements:
        - name: "ALL"
          description: "Get all phone Number type"
        - name: "Voice"
          description: "Get only voice number type"
        - name: "SMS"
          description: "Get Only SMs number type"
      - name: "FriendlyName"
        in: "formData"
        description: "A human-readable label added to the number object."
        required: false
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.phone_number_controller"
  /incomingphone/massupdatenumber.json:
    post:
      tags:
      - "Phone Number"
      summary: "Bulk Update Numbers"
      description: "Update properties for a Ytel numbers that has been purchased for\
        \ your account. Refer to the parameters list for the list of properties that\
        \ can be updated."
      operationId: "incomingphone_massupdatenumber"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "PhoneNumber"
        in: "formData"
        description: "A valid comma(,) separated Ytel numbers. (E.164 format)."
        required: true
        type: "string"
      - name: "VoiceUrl"
        in: "formData"
        description: "The URL returning InboundXML incoming calls should execute when\
          \ connected."
        required: true
        type: "string"
      - name: "FriendlyName"
        in: "formData"
        description: "A human-readable value for labeling the number."
        required: false
        type: "string"
      - name: "VoiceMethod"
        in: "formData"
        description: "Specifies the HTTP method used to request the VoiceUrl once\
          \ incoming call connects."
        required: false
        type: "string"
      - name: "VoiceFallbackUrl"
        in: "formData"
        description: "URL used if any errors occur during execution of InboundXML\
          \ on a call or at initial request of the voice url"
        required: false
        type: "string"
      - name: "VoiceFallbackMethod"
        in: "formData"
        description: "Specifies the HTTP method used to request the VoiceFallbackUrl\
          \ once incoming call connects."
        required: false
        type: "string"
      - name: "HangupCallback"
        in: "formData"
        description: "URL that can be requested to receive notification when and how\
          \ incoming call has ended."
        required: false
        type: "string"
      - name: "HangupCallbackMethod"
        in: "formData"
        description: "The HTTP method Ytel will use when requesting the HangupCallback\
          \ URL."
        required: false
        type: "string"
      - name: "HeartbeatUrl"
        in: "formData"
        description: "URL that can be used to monitor the phone number."
        required: false
        type: "string"
      - name: "HeartbeatMethod"
        in: "formData"
        description: "The HTTP method Ytel will use when requesting the HeartbeatUrl."
        required: false
        type: "string"
      - name: "SmsUrl"
        in: "formData"
        description: "URL requested when an SMS is received."
        required: false
        type: "string"
      - name: "SmsMethod"
        in: "formData"
        description: "The HTTP method Ytel will use when requesting the SmsUrl."
        required: false
        type: "string"
      - name: "SmsFallbackUrl"
        in: "formData"
        description: "URL used if any errors occur during execution of InboundXML\
          \ from an SMS or at initial request of the SmsUrl."
        required: false
        type: "string"
      - name: "SmsFallbackMethod"
        in: "formData"
        description: "The HTTP method Ytel will use when URL requested if the SmsUrl\
          \ is not available."
        required: false
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.phone_number_controller"
  /incomingphone/getdidscore.json:
    post:
      tags:
      - "Phone Number"
      summary: "Get DID Score"
      description: "Get DID Score Number"
      operationId: "incomingphone_getdidscore"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "Phonenumber"
        in: "formData"
        description: "Specifies the multiple phone numbers for check updated spamscore\
          \ ."
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.phone_number_controller"
  /incomingphone/bulkbuy.json:
    post:
      tags:
      - "Phone Number"
      summary: "Bulk Buy Numbers"
      description: "Purchase a selected number of DID's from specific area codes to\
        \ be used with your Ytel account."
      operationId: "incomingphone_bulkbuy"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "NumberType"
        in: "formData"
        description: "The capability the number supports."
        required: true
        type: "string"
        enum:
        - "all"
        - "voice"
        - "sms"
        x-enum-elements:
        - name: "ALL"
          description: "Get all phone Number type"
        - name: "Voice"
          description: "Get only voice number type"
        - name: "SMS"
          description: "Get Only SMs number type"
      - name: "AreaCode"
        in: "formData"
        description: "Specifies the area code for the returned list of available numbers.\
          \ Only available for North American numbers."
        required: true
        type: "string"
      - name: "Quantity"
        in: "formData"
        description: "A positive integer that tells how many number you want to buy\
          \ at a time."
        required: true
        type: "string"
      - name: "Leftover"
        in: "formData"
        description: "If desired quantity is unavailable purchase what is available\
          \ ."
        required: false
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.phone_number_controller"
...
securityDefinitions:
  auth:
    type: "basic"
    x-testValue: []
    x-skip-client-authentication: false
...