# rails5/app/controllers/phone_number_controller.rb
=begin
Ytel API V3

Ytel API version 3

OpenAPI spec version: 3.12

Generated by: https://github.com/swagger-api/swagger-codegen.git

=end
class PhoneNumberController < ApplicationController

  def incomingphone_availablenumber
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def incomingphone_bulkbuy
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def incomingphone_buynumber
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def incomingphone_getdidscore
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def incomingphone_listnumber
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def incomingphone_massreleasenumber
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def incomingphone_massupdatenumber
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def incomingphone_releasenumber_by_response_type_post
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def incomingphone_transferphonenumbers
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def incomingphone_updatenumber
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def incomingphone_viewnumber
    # Your code here

    render json: {"message" => "yes, it worked"}
  end
end